+++ Runtime/Abstract/UnityEngine/WSA/Cursor.cs
@@ -1,0 +1,14 @@
+namespace Moqunity.Abstract.UnityEngine.WSA
+{
+	/// <summary>
+	///   <para>Cursor API for Windows Store Apps.</para>
+	/// </summary>
+	public interface Cursor
+	{
+		/// <summary>
+		///   <para>Set a custom cursor.</para>
+		/// </summary>
+		/// <param name="id">The cursor resource id.</param>
+		void SetCustomCursor(uint id);
+	}
+}