+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Broken test assembly
 ### Added
 - Random interface, implementation, base fake
+- WSA.Application interface
+- WSA.Cursor interface
+- WSA.Launcher interface
+- Rendering.AsyncGPUReadback interface
 
 ## [0.2.1] - 2022-04-06
 ### Fixed